+++ src/lib/components/FormRenderer.tsx
@@ -1,34 +1,49 @@
-import React from 'react'
-import { Formik, Form, Field } from 'formik'
+import React, { useRef, useEffect } from 'react'
+import { Formik, Form, Field, useFormikContext, FormikProps } from 'formik'
 import * as yup from 'yup'
 import { useDndEditorContext } from '../DndEditorProvider'
 import Container from '../assets/Container'
 import { DndComponentItem } from '../types'
+import { checkForDiplayCondition } from '../utils'
+import FormObserver from './preferences/components/FormObserver'
 
 export interface FormRendererProps {
     onSubmit: (value: any) => void
+    onChange: (value: any) => void
 }
 
-const FormRenderer: React.FC<FormRendererProps> = ({ onSubmit }) => {
+const FormRenderer: React.FC<FormRendererProps> = ({ onSubmit, onChange, ...props }) => {
     const renderProps = useDndEditorContext()
 
-    const Children = () => (
-        <div>
-            {renderProps.state.items?.map((item) => {
-                const stateItem = renderProps.state.entities[item.id]
-                const name = stateItem.name
-                const updatedRenderProps = { ...renderProps, item, name }
-                return (
-                    <div key={item.id}>
-                        {Container.render(
-                            updatedRenderProps,
-                            renderProps.itemsMap[stateItem.parent.id]?.render?.(updatedRenderProps)
-                        )}
-                    </div>
-                )
-            })}
-        </div>
-    )
+    const Children = () => {
+        const formik = useFormikContext()
+        return (
+            <div>
+                {renderProps.state.items
+                    ?.filter((item) =>
+                        checkForDiplayCondition(
+                            renderProps.state.entities[item.id].values.__condition,
+                            formik
+                        )
+                    )
+                    .map((item) => {
+                        const stateItem = renderProps.state.entities[item.id]
+                        const name = stateItem.name
+                        const updatedRenderProps = { ...renderProps, item, name }
+                        return (
+                            <div key={item.id}>
+                                {Container.render(
+                                    updatedRenderProps,
+                                    renderProps.itemsMap[stateItem.parent.id]?.render?.(
+                                        updatedRenderProps
+                                    )
+                                )}
+                            </div>
+                        )
+                    })}
+            </div>
+        )
+    }
 
     let validation: any = {}
     renderProps.state.items?.map((item) => {
@@ -42,16 +57,13 @@ const FormRenderer: React.FC<FormRendererProps> = ({ onSubmit }) => {
     })
     const validationSchema = yup.object().shape(validation)
 
+    //@ts-ignore
     return (
-        <Formik
-            enableReinitialize
-            initialValues={{}}
-            onSubmit={(s) => onSubmit && onSubmit(s)}
-            validationSchema={validationSchema}
-        >
+        <Formik enableReinitialize initialValues={{}} onSubmit={(s) => onSubmit && onSubmit(s)}>
             <Form style={{ paddingBottom: '50px' }}>
                 {renderProps.template.render(renderProps, <Children />)}
-                <Field type="submit" value="Submit" id="submit"></Field>
+                {props.children}
+                <FormObserver onChange={onChange} />
             </Form>
         </Formik>
     )