+++ src/lib/components/preferences/items/InputOptions.tsx
@@ -2,16 +2,31 @@ import { Trans } from '@lingui/macro'
 import React from 'react'
 import { InputProps } from '@material-ui/core'
 import LabeledTextInput from './LabeledTextInput'
+import { StringFormValue } from '../../../types'
 
-export interface OptionsProps extends Omit<InputProps, 'value' | 'onChange'> {
-    value: string[]
-    onChange: (value: string[]) => void
+export type InputOption = {
+    key: string
+    label: string
+    value: StringFormValue
 }
-const InputOptions: React.FC<OptionsProps> = ({ onChange, value, ...props }) => {
-    const handleOnChange = (value: string) => {
-        onChange(value.split('\n'))
+export interface OptionsProps extends Omit<InputProps, 'value' | 'onChange'> {
+    value: InputOption[]
+    onChange: (value: InputOption[]) => void
+}
+const InputOptions: React.FC<OptionsProps> = ({ onChange, value }) => {
+    const handleOnChange = (value: any) => {
+        onChange(
+            value.split('\n').map((option: string) => ({
+                key: option.split(' ').join('-').toLowerCase(),
+                label: option,
+                value: {
+                    text: option,
+                    valueType: 'String'
+                }
+            }))
+        )
     }
-    const inputValue = value && value.join('\n')
+    const inputValue = value && value.map((v) => v.label).join('\n')
     return (
         <LabeledTextInput
             label={<Trans>Options</Trans>}