+++ src/lib/index.ts
@@ -1,2 +1,4 @@
 export { default as DndEditor } from './DndEditor'
 export * from './DndEditor'
+export { default as Renderer } from './Renderer'
+export * from './Renderer'