+++ src/lib/types.ts
@@ -129,6 +129,7 @@ export type DndBlockItem = DndBaseItem & {
     parent?: string
     image: string
     icon?: React.ComponentType<SvgIconProps>
+    generateSettings?: (renderProps: RenderProps) => DndItemSetting[]
 }
 export type DndItem = DndGroupItem | DndBlockItem
 
@@ -159,6 +160,28 @@ export interface DndEditorContextProps {
     buildermode: boolean
 }
 
+export type BooleanFormValue = {
+    boolean: boolean
+    valueType: 'Boolean'
+}
+
+export type StringFormValue = {
+    text: string
+    valueType: 'String'
+}
+
+export type DecimalFormValue = {
+    decimal: number
+    valueType: 'Decimal'
+}
+
+export type NumberFormValue = {
+    number: number
+    valueType: 'Number'
+}
+
+export type FormValue = BooleanFormValue | StringFormValue | DecimalFormValue | NumberFormValue
+
 export interface BlockRendererProps {
     settings: DndItemSetting[]
     state: InitialValues
@@ -167,4 +190,5 @@ export interface BlockRendererProps {
 
 export type RenderProps = DndEditorContextProps & {
     item?: DndStateItemEntity
+    name?: string
 }