+++ src/stories/Editor.stories.tsx
@@ -1,20 +1,28 @@
-import React from 'react';
-import { Story, Meta } from '@storybook/react/types-6-0';
-import { ThemeProvider } from '@material-ui/core';
+import React from 'react'
+import { Story, Meta } from '@storybook/react/types-6-0'
+import { ThemeProvider } from '@material-ui/core'
 import { createMuiTheme } from '@material-ui/core/styles'
-import { createDndState } from '../lib/utils';
-import { DndEditor, DndEditorProps } from '../lib';
+import { createDndState } from '../lib/utils'
+import { DndEditor, DndEditorProps } from '../lib'
 import * as Blocks from '../lib/assets/blocks'
 import * as Groups from '../lib/assets/groups'
 import * as Templates from '../lib/assets/templates'
-import { DndItem } from '../lib/types';
-import FormElements from '../lib/assets/groups/FormElements';
+import { DndItem } from '../lib/types'
+import FormElements from '../lib/assets/groups/FormElements'
 
 export default {
-  title: 'Editor',
-  component: DndEditor,
-  decorators: [story => <ThemeProvider theme={createMuiTheme({ typography: { fontFamily: '"Poppins", sans-serif' } })}>{story()}</ThemeProvider>]
-} as Meta;
+    title: 'Editor',
+    component: DndEditor,
+    decorators: [
+        (story) => (
+            <ThemeProvider
+                theme={createMuiTheme({ typography: { fontFamily: '"Poppins", sans-serif' } })}
+            >
+                {story()}
+            </ThemeProvider>
+        )
+    ]
+} as Meta
 
 const defaultProps = {
     smartyTags: {
@@ -27,20 +35,21 @@ const defaultProps = {
     },
     value: createDndState({}),
     items: [...Object.values(Groups), ...Object.values(Blocks)],
-    onChange: (_: any) => {},
+    onChange: (_: any) => {}
 }
 
-const Template: Story<DndEditorProps> = (args) => <DndEditor {...args} />;
+const Template: Story<DndEditorProps> = (args) => <DndEditor {...args} />
 
-export const Mail = Template.bind({});
+export const Mail = Template.bind({})
 Mail.args = {
-    ...defaultProps, template : Templates.Mail
-};
+    ...defaultProps,
+    template: Templates.Mail
+}
 
-const formItems: DndItem[] = [
-    FormElements, ...Object.values(Blocks)
-]
-export const Form = Template.bind({});
+const formItems: DndItem[] = [FormElements, ...Object.values(Blocks)]
+export const Form = Template.bind({})
 Form.args = {
-    ...defaultProps, items: formItems, template : Templates.Form
-};
+    ...defaultProps,
+    items: formItems,
+    template: Templates.Form
+}